requirements.txt
openai
tiktoken

langchain==0.2.14
langchain-openai==0.1.22
langchain-community==0.2.12
langchain-text-splitters==0.2.2
langchain-core==0.2.43

pymilvus==2.6.9
protobuf==6.33.5
grpcio==1.78.0
grpcio-tools==1.78.0

SQLAlchemy
python-dotenv

scikit-learn

fastapi
uvicorn

langgraph==0.2.36